fix: change the length of the unit of work in order to fit the transaction
Unit of work is an important concept in SQLAlchemy. Previously, we were attempting to do two different transactions in the same unit of work in a session and this was leading to an undefined state of the session adapter. Now, we insert and then select in the same transaction instead of commiting and then trying to select again leading to a misinterpretation of the state of the database